requirements.txt
backports.functools-lru-cache==1.5
cachetools==2.0.1
click==6.7
cycler==0.10.0
Django==1.11.4
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
joblib==0.11
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
nltk==3.2.5
numpy==1.13.1
pandas==0.21.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
redis==2.10.6
scikit-learn==0.19.1
scikit-surprise==1.0.3
scipy==1.0.0
six==1.10.0
Unidecode==0.4.21
Werkzeug==0.13
django-filter==1.0.4
django-material==1.2.5
django-viewflow==1.0.2
gunicorn

